Struct ink_model::ExecutionEnv

pub struct ExecutionEnv<State, Env> {
    pub state: State,
    // some fields omitted
}

Provides a safe interface to an environment given a contract state.

Fields
state: State

The contract state.

Methods

impl<State, Env> ExecutionEnv<State, Env>

pub fn split(&self) -> (&EnvHandler<Env>, &State)

Splits the execution environment into shared references to the environment handler and the state.

Note

This can be useful if you want to implement a message as a method of the state to make it callable from other messages.

pub fn split_mut(&mut self) -> (&mut EnvHandler<Env>, &mut State)

Splits the execution environment into mutable references to the environment handler and the state.

Note

This can be useful if you want to implement a message as a method of the state to make it callable from other messages.