Trait ink_model::Contract

pub trait Contract {
    fn deploy(self);

    fn dispatch(self);
}

A simple interface to work with contracts.

Required methods

fn deploy(self)

Deploys the contract state.

Should be performed exactly once during contract lifetime. Consumes the contract since nothing should be done afterwards.

fn dispatch(self)

Dispatches the call input to a pre defined contract message and runs its handler.

Consumes self since it should be the default action after instantiation.

Panics

Panics (Wasm: traps) if the call input was invalid. The call input is invalid if there was no matching function selector found or if the data for a given selected function was not decodable.

Loading content...