Trait ink_model::TestableContract

pub trait TestableContract {
    type DeployArgs: Encode;
    fn deploy(&mut self, deploy_args: Self::DeployArgs);

    fn call<Msg>(
        &mut self, 
        input: <Msg as Message>::Input
    ) -> <Msg as Message>::Output
    where
        Msg: Message,
        <Msg as Message>::Input: Encode,
        <Msg as Message>::Output: Decode;
}

An interface that allows for simple testing of contracts.

Associated Types

type DeployArgs: Encode

The arguments used for deployment.

These must be the same as the ones defined on the deploy handler of a contract declaration.

Loading content...
Required methods

fn deploy(&mut self, deploy_args: Self::DeployArgs)

Deploys the contract given the provided arguments for deployment.

Note

This shall be performed only once during the lifetime of a contract.

Panics

This might panic if the provided arguments do not match the expected.

fn call<Msg>(
&mut self,
input: <Msg as Message>::Input
) -> <Msg as Message>::Output where
Msg: Message,
<Msg as Message>::Input: Encode,
<Msg as Message>::Output: Decode,

Calls the contract with the given message and its inputs and upon successful execution returns its result.

Note

Takes &mut self since it could potentially call a message that mutates state. There currently is no separation between messages that mutate state and those that do not.

Panics

If the contract has no message handler setup for the given message.

Loading content...